Fix issue on Decoding image for AI Builder
Issue about decoding image when the FLOW is submitted in the Power Automate Mobile Application. You will need to use the function "base64ToBinary" function in the Predict Part. It will not works for both (Desktop in the Web Client of PowerAutomate and Mobile App)
Like this : https://imgshare.io/image/N6hlut